Rename NodeServerStub to TransactionServiceStub
